Fix notifier and step descriptors; rename plan to a cycle
