+++ src/main/resources/com/testquality/jenkins/TestQualityStep/config.jelly
@@ -1,13 +1,13 @@
 <?jelly escape-by-default='true'?>
 <j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
     <f:entry title="Project" field="project">
-        <f:select />
+        <f:textbox />
     </f:entry>
-    <f:entry title="Test Plan" field="plan">
-        <f:select />
+    <f:entry title="Cycle" field="cycle">
+        <f:textbox />
     </f:entry>
     <f:entry title="Milestone" field="milestone">
-        <f:select />
+        <f:textbox />
     </f:entry>
     <f:entry title="Test Results XMLs" field="testResults">
         <f:textbox />
