+++ src/main/resources/com/testquality/jenkins/TestQualityStep/help-cycle.html
@@ -1,0 +1,4 @@
+<!--Used in snippet generation-->
+<div>
+  Optional field. Cycle will be modified to reflect test suites and test cases in test results. One test run will be added for each job execution.
+</div>
