+++ src/main/resources/com/testquality/jenkins/TestQualityStep/help-milestone.html
@@ -1,0 +1,4 @@
+<!--Used in snippet generation-->
+<div>
+  Milestone is optional however if used the test run will be associated with the approperate milestone.
+</div>
