+++ src/main/resources/com/testquality/jenkins/TestQualityStep/help-project.html
@@ -1,0 +1,4 @@
+<!--Used in snippet generation-->
+<div>
+  Select a valid TestQuality project, based on connection provided in Configure System.
+</div>
