+++ src/main/resources/com/testquality/jenkins/TestQualityStep/help-testResults.html
@@ -1,0 +1,4 @@
+<!--Used in snippet generation-->
+<div>
+  Provide valid pattern to test result xml files. Wildcards "*" can be used in name and "**" can be used to traverse directories.
+</div>
